requirements.txt
python-dotenv==1.0.0
pyodbc==5.0.1
psycopg2-binary==2.9.9